# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "properties": {
          "code": {
            "type": "string"
          },
          "details": true,
          "error": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "details",
          "error"
        ],
        "title": "ApiErrorBody",
        "type": "object"
      },
...
      "PeopleListResponse": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "properties": {
          "people": {
            "items": {
              "$ref": "#/components/schemas/PersonProfile"
            },
            "type": "array"
          }
        },
        "required": [
          "people"
        ],
        "title": "PeopleListResponse",
        "type": "object"
      },
      "PersonCreateRequest": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "properties": {
          "discord_id": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "discord_username": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "feishu_id": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "profile_content": {
            "default": "",
            "type": "string"
          },
          "telegram_id": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "name"
        ],
        "title": "PersonCreateRequest",
        "type": "object"
      },
      "PersonMutationResponse": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "ok": {
            "type": "boolean"
          },
          "person": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonProfile"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "ok"
        ],
        "title": "PersonMutationResponse",
        "type": "object"
      },
      "PersonProfile": {
        "properties": {
          "created_at": {
            "format": "int64",
            "type": "integer"
          },
          "discord_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "discord_username": {
            "type": [
              "string",
              "null"
            ]
          },
          "feishu_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "profile_content": {
            "type": "string"
          },
          "telegram_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "format": "int64",
            "type": "integer"
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "created_at",
          "id",
          "name",
          "profile_content",
          "updated_at"
        ],
        "type": "object"
      },
      "PersonUpdateRequest": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "properties": {
          "discord_id": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "discord_username": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "feishu_id": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "profile_content": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "telegram_id": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          }
        },
        "title": "PersonUpdateRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "allthecodes Web API",
    "version": "v1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/people": {
      "get": {
        "operationId": "PeopleList",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeopleListResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "concurrent"
      },
      "post": {
        "operationId": "PeopleCreate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMutationResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "concurrent"
      }
    },
    "/api/people/{id}": {
      "delete": {
        "operationId": "PeopleDelete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMutationResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "concurrent"
      },
      "get": {
        "operationId": "PeopleDetail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMutationResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "concurrent"
      },
      "patch": {
        "operationId": "PeopleUpdate",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMutationResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "concurrent"
      }
    },
...
  }
}